:recycle: refactor: SettingControl's logic splited
